+++ NEWS.md
@@ -1,5 +1,6 @@
 ### crunch 1.15.1 (under development)
 * Support for additional dataset export arguments
+* Add `is.derived` method for Variables
 
 ## crunch 1.15.0
 * `multitables` method to access catalog from a Dataset. `newMultitable` to create one. See `?multitables` and `?newMultitable` for docs and examples.